requirements.txt
boto3==1.4.2
botocore==1.4.85
docutils==0.13.1
futures==3.0.5
jmespath==0.9.0
pkg-resources==0.0.0
python-dateutil==2.6.0
s3transfer==0.1.9
six==1.10.0
